Extract enums for orders and payments
