+++ saleor/graphql/order/resolvers.py
@@ -5,7 +5,8 @@ from ...order import OrderEvents, OrderStatus, models
 from ...order.utils import sum_order_totals
 from ...shipping import models as shipping_models
 from ..utils import filter_by_period, filter_by_query_param
-from .types import Order, OrderStatusFilter
+from .enums import OrderStatusFilter
+from .types import Order
 
 ORDER_SEARCH_FIELDS = (
     'id', 'discount_name', 'token', 'user_email', 'user__email')
