+++ saleor/graphql/order/schema.py
@@ -7,6 +7,7 @@ from ..core.fields import PrefetchingConnectionField
 from ..core.enums import ReportingPeriod
 from ..core.types import TaxedMoney
 from ..descriptions import DESCRIPTIONS
+from .enums import OrderStatusFilter
 from .mutations.draft_orders import (
     DraftOrderComplete, DraftOrderCreate, DraftOrderDelete,
     DraftOrderLineCreate, DraftOrderLineDelete, DraftOrderLineUpdate,
@@ -19,7 +20,7 @@ from .mutations.orders import (
 from .resolvers import (
     resolve_homepage_events, resolve_order, resolve_orders,
     resolve_orders_total)
-from .types import Order, OrderEvent, OrderStatusFilter
+from .types import Order, OrderEvent
 
 
 class OrderQueries(graphene.ObjectType):
