+++ saleor/graphql/order/types.py
@@ -2,22 +2,15 @@ import graphene
 import graphene_django_optimizer as gql_optimizer
 from graphene import relay
 
-from ...order import OrderEvents, OrderEventsEmails, models
+from ...order import models
 from ...product.templatetags.product_images import get_product_image_thumbnail
 from ..account.types import User
 from ..core.connection import CountableDjangoObjectType
 from ..core.types.money import Money, TaxedMoney
 from ..payment.types import OrderAction, Payment, PaymentChargeStatusEnum
 from ..shipping.types import ShippingMethod
+from .enums import OrderEventsEnum, OrderEventsEmailsEnum
 from .utils import can_finalize_draft_order
-
-OrderEventsEnum = graphene.Enum.from_enum(OrderEvents)
-OrderEventsEmailsEnum = graphene.Enum.from_enum(OrderEventsEmails)
-
-
-class OrderStatusFilter(graphene.Enum):
-    READY_TO_FULFILL = 'READY_TO_FULFILL'
-    READY_TO_CAPTURE = 'READY_TO_CAPTURE'
 
 
 class OrderEvent(CountableDjangoObjectType):
