+++ saleor/graphql/payment/types.py
@@ -3,35 +3,11 @@ import graphene_django_optimizer as gql_optimizer
 from django_countries.fields import Country
 from graphene import relay
 
-from ...payment import GATEWAYS_ENUM, ChargeStatus, models
+from ...payment import models
 from ..account.types import Address
 from ..core.connection import CountableDjangoObjectType
 from ..core.types import Money
-from ..core.utils import str_to_enum
-
-PaymentGatewayEnum = graphene.Enum.from_enum(GATEWAYS_ENUM)
-PaymentChargeStatusEnum = graphene.Enum(
-    'PaymentChargeStatusEnum',
-    [(str_to_enum(code.upper()), code) for code, name in ChargeStatus.CHOICES])
-
-
-class OrderAction(graphene.Enum):
-    # FIXME could use a better name
-    CAPTURE = 'CAPTURE'
-    MARK_AS_PAID = 'MARK_AS_PAID'
-    REFUND = 'REFUND'
-    VOID = 'VOID'
-
-    @property
-    def description(self):
-        if self == OrderAction.CAPTURE:
-            return 'Represents the capture action.'
-        if self == OrderAction.MARK_AS_PAID:
-            return 'Represents a mark-as-paid action.'
-        if self == OrderAction.REFUND:
-            return 'Represents a refund action.'
-        if self == OrderAction.VOID:
-            return 'Represents a void action.'
+from .enums import OrderAction, PaymentChargeStatusEnum
 
 
 class Transaction(CountableDjangoObjectType):
@@ -45,7 +21,6 @@ class Transaction(CountableDjangoObjectType):
         filter_fields = ['id']
         exclude_fields = ['currency']
 
-    @staticmethod
     def resolve_amount(self, info):
         return self.get_amount()
 
@@ -68,9 +43,6 @@ class CreditCard(graphene.ObjectType):
 
 
 class Payment(CountableDjangoObjectType):
-    # FIXME gateway_response field should be resolved
-    # if we want to use it on the frontend
-    # otherwise it should be removed from this type
     charge_status = PaymentChargeStatusEnum(
         description='Internal payment status.', required=True)
     actions = graphene.List(
